src/sentinel/requirements.txt
functions-framework==3.*
google-cloud-pubsub
google-cloud-firestore
google-cloud-logging
requests
gspread
google-auth